+++ Modules/MetalModule/Sources/MetalModule/Factory/MetalModuleFactory.swift
@@ -1,0 +1,13 @@
+//
+//  MetalModuleFactory.swift
+//  MetalModule
+//
+//  Created by max on 24.05.2026.
+//
+
+@MainActor
+public enum MetalModuleFactory {
+    public static func makeResources() -> MetalModuleResources {
+        MetalModuleResources()
+    }
+}
