+++ Modules/MetalModule/Sources/MetalModule/Factory/MetalModuleResources/MetalModuleResources+MetalModuleTransferOrbitControlling.swift
@@ -1,0 +1,12 @@
+//
+//  MetalModuleResources+MetalModuleTransferOrbitControlling.swift
+//  MetalModule
+//
+//  Created by max on 24.05.2026.
+//
+
+extension MetalModuleResources: MetalModuleTransferOrbitControlling {
+    public func showTransferOrbit(to destinationName: String) {
+        renderer?.showTransferOrbit(to: destinationName)
+    }
+}
