+++ Modules/MetalModule/Sources/MetalModule/Factory/MetalModuleTransferOrbitControlling.swift
@@ -1,0 +1,11 @@
+//
+//  MetalModuleTransferOrbitControlling.swift
+//  MetalModule
+//
+//  Created by max on 24.05.2026.
+//
+
+@MainActor
+public protocol MetalModuleTransferOrbitControlling: AnyObject {
+    func showTransferOrbit(to destinationName: String)
+}
